Removed old lyapunov memoizer and system size reduction
