+++ src/qttools/boundary_conditions/lyapunov_solver/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2024-2026 ETH Zurich and the authors of the qttools package.
 
-from qttools.lyapunov.doubling import Doubling
-from qttools.lyapunov.lyapunov import LyapunovMemoizer, LyapunovSolver
-from qttools.lyapunov.spectral import Spectral
+from qttools.boundary_conditions.lyapunov_solver.doubling import Doubling
+from qttools.boundary_conditions.lyapunov_solver.spectral import Spectral
 
-__all__ = ["LyapunovSolver", "LyapunovMemoizer", "Doubling", "Spectral"]
+__all__ = ["Doubling", "Spectral"]
